### 2019-08-05 #333 @nevrome
Submission: c14bazAAR
Submitting Author: Clemens Schmid (@nevrome)
Repository: https://github.com/ISAAKiel/c14bazAAR
Version submitted: 1.0.3.9000
Version accepted: TBD

---

```
Package: c14bazAAR
Title: Download and Prepare C14 Dates from Different Source Databases
Description: Query different C14 date databases and apply basic data cleaning, merging and calibration steps.
Version: 1.0.3.9000
    c(person(given = "Clemens",
             family = "Schmid",
             email = "clemens@nevrome.de",
      person(given = "Dirk",
             family = "Seidensticker",
             role = "aut",
             email = "dirk.seidensticker@gmail.com",
      person(given = "Daniel",
             family = "Knitter",
             role = "aut",
             email = "knitter@geographie.uni-kiel.de",
      person(given = "Martin",
             family = "Hinz",
             role = "aut",
             email = "martin.hinz@iaw.unibe.ch",
      person(given = "David",
             family = "Matzig",
             role = "aut",
             email = "matzigdavid@gmail.com",
      person(given = "Wolfgang",
             family = "Hamer",
             role = "aut",
             email = "hamer@geographie.uni-kiel.de",
      person(given = "Kay",
             family = "Schmütz",
             role = "aut",
             email = "kschmuetz@ufg.uni-kiel.de"),
      person(given = "Nils",
             family = "Mueller-Scheessel",
             role = "ctb",
             email = "nils.mueller-scheessel@ufg.uni-kiel.de",
             comment = c(ORCID = "0000-0001-7992-8722")))
Language: en_GB
License: GPL-2 | file LICENSE
    crayon (>= 1.3.4),
    data.table (>= 1.11.4),
    magrittr (>= 1.5),
    pbapply (>= 1.3-3),
    rlang (>= 0.1.1),
    stats (>= 3.4.0),
    tibble (>= 1.3.3),
    tidyr (>= 0.6.3)
    Bchron,
    countrycode,
    dataverse,
    ggridges,
    lwgeom,
    mapview,
    openxlsx,
    plyr,
    rgeos,
    rnaturalearth,
    rworldmap,
    rworldxtra,
    sf,
    stringdist,

```

c14bazAAR was created to access radiocarbon dates from openly accessible archives. All functions are related to data download and data preparation.

Mostly archaeologists as most radiocarbon databases currently accessible via c14bazAAR contain radiocarbon dates from archaeological contexts. The package could also be of interest for geographers and other kinds of ecologists -- especially if non-archaeological data sources are added to the portfolio.

No. Not that I'm aware of.

It already is on CRAN in version [1.0.3](https://CRAN.R-project.org/package=c14bazAAR).

  - [x] The package is deposited in a long-term repository with the DOI: 10.17605/OSF.IO/3DS6A
