### 2019-08-06 #333 @melvidoni
Hello @nevrome and thank you for your submission. I have noticed that you haven't ticked the rOpenSci's CoC box. Is that for a particular reason? It is fundamental for us to have that, otherwise we cannot continue the process.
