### 2019-08-06 #333 @melvidoni
### Editor checks:

---

#### Editor comments

Thank you for submitting, I'll be your handling editor. I run `goodpractices` but apparently you are clear to go, so I will start finding reviewers for the package. I'll let you know once they are in, and we have a review deadline.

---

Reviewers: TBD
Due date: TBD
