### 2019-09-05 #333 @benmarwick
- [x] As the reviewer I confirm that there are no [conflicts of interest](#coi) for me to review this work

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is an exemplary package that was a delight to review. It is a perfect fit for rOpenSci and deserves a wide audidence.

I ran some of the code in the README.md and everything worked as expected. I appreciated the comments that some code would take long to run! The vignette is beautifully written and elegantly and efficiently demonstrates some typical uses of the package. The function help files are thorough and encyclopedic, and rich with examples. All of the metadata files are excellent and have been prepared to the highest standards. I inspected a sample of functions and everything I saw looked excellent.

One suggestion for improvement: in `get_country_thesaurus` and similar `get_` functions I see we are downloading data, eg. from `"https://raw.githubusercontent.com/ISAAKiel/c14bazAAR/master/data-raw/material_thesaurus.csv"`. Looking in `data-raw/` I see that these files are not large (none more than 100 KB). So I suggest that the authors make these exported data objects in the package so users don't have to download again.

I used [pkgreviewr](https://ropenscilabs.github.io/pkgreviewr/index.html]) to help prepare this review, the full contents of my review are in the fold below

<details><summary>pkgreviewr output</summary>

---
    html_notebook:
        toc: true
        toc_float: true
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  fig.path = "man/figures/README-",
  out.width = "100%"
)

library(magrittr)
library(devtools)
```

# `c14bazAAR` - package review

## **Reviewer:** [\@benmarwick](https://github.com/benmarwick)

### Review Submitted:
**`r cat(sprintf("**Last updated:** %s", Sys.Date()))`**

***

This report contains documents associated with the review of **rOpenSci** submitted package:

## Package info

**Description:**

Query different C14 date databases and apply basic data cleaning, merging and calibration steps.

**Author:** `r
    c(person(given = "Clemens",
             family = "Schmid",
             email = "clemens@nevrome.de",
      person(given = "Dirk",
             family = "Seidensticker",
             role = "aut",
             email = "dirk.seidensticker@gmail.com",
      person(given = "Daniel",
             family = "Knitter",
             role = "aut",
             email = "knitter@geographie.uni-kiel.de",
      person(given = "Martin",
             family = "Hinz",
             role = "aut",
             email = "martin.hinz@iaw.unibe.ch",
      person(given = "David",
             family = "Matzig",
             role = "aut",
             email = "matzigdavid@gmail.com",
      person(given = "Wolfgang",
             family = "Hamer",
             role = "aut",
             email = "hamer@geographie.uni-kiel.de",
      person(given = "Kay",
             family = "Schmütz",
             role = "aut",
             email = "kschmuetz@ufg.uni-kiel.de"),
      person(given = "Nils",
             family = "Mueller-Scheessel",
             role = "ctb",
             email = "nils.mueller-scheessel@ufg.uni-kiel.de",
             comment = c(ORCID = "0000-0001-7992-8722")))`

**repo url:** <https://github.com/ISAAKiel/c14bazAAR>

**website url:** <>

## Review info

#### See [reviewer guidelines](https://ropensci.github.io/dev_guide/reviewerguide.html) for further information on the rOpenSci review process.

**key review checks:**

- Does the code comply with **general principles in the [Mozilla reviewing guide](https://mozillascience.github.io/codeReview/review.html)**?
- Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?
- Are there **improvements** that could be made to the **code style?**
- Is there **code duplication** in the package that should be reduced?
- Are there **user interface improvements** that could be made?
- Are there **performance improvements** that could be made?
- Is the [**documentation**](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) **clear and sufficient**?

Please be respectful and kind to the authors in your reviews. The rOpenSci [code of conduct](https://ropensci.github.io/dev_guide/policies.html#code-of-conduct) is mandatory for everyone involved in our review process.

***

### session info

```{r pkg_dir, echo = F}
pkg_dir <- "/Users/bmarwick/Desktop/c14bazAAR-review/c14bazAAR"
```

## Test installation

### test local `c14bazAAR` install:

```{r test-local}
install(pkg_dir, dependencies = T, build_vignettes = T)
```

```{r github-rm}
remove.packages("c14bazAAR")
```
#### **comments:**

No issues with local install of `c14bazAAR`.

***

### test install of `c14bazAAR` from GitHub with:

```{r test-github}
devtools::install_github("ISAAKiel/c14bazAAR", dependencies = T, build_vignettes = T)
```

#### **comments:**

No issues with GitHub install of `c14bazAAR`

***

## Check package integrity

### run checks on `c14bazAAR` source:

```{r check-checks}
devtools::check(pkg_dir)
```
#### **comments:**

No issues with checks of `c14bazAAR`:

```
── R CMD check results ─────────────────── c14bazAAR 1.0.3.9000 ────
Duration: 32.6s

0 errors ✔ | 0 warnings ✔ | 0 notes ✔
```

***

### run tests on `c14bazAAR` source:

```{r check-tests}
devtools::test(pkg_dir)
```
#### **comments:**

No issues with tests of `c14bazAAR`:

```
══ Results ════════════════════

OK:       68
```

***

### check `c14bazAAR` for goodpractice:

```{r test-goodpractice}
goodpractice::gp(pkg_dir)
```
#### **comments:**

Many warnings of 'write unit tests for all functions, and all package code in general. 42% of code lines are covered by test cases.' Looks like these warning are appearing on unexported helper functions that do checking and formatting. I don't believe it is a high priority to have test coverage on these.

***

## Check package metadata files

### inspect

- #### [README](https://github.com/ISAAKiel/c14bazAAR)
- #### [DESCRIPTION](https://github.com/ISAAKiel/c14bazAAR/blob/master/DESCRIPTION)
- #### [NAMESPACE](https://github.com/ISAAKiel/c14bazAAR/blob/master/NAMESPACE)

### spell check

```{r spell-check}
devtools::spell_check(pkg_dir)
```

#### **comments:**

All of these metadata files are excellent and have been prepared to the highest standards.

***

## Check documentation

online documentation: **<>**

* Is the [documentation](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) clear and sufficient?

### test `c14bazAAR` function help files:

```{r test-help}
help(package = "c14bazAAR")
```

#### **comments:**

The function help files are thorough and encyclopeadic, and rich with examples.

***

### test `c14bazAAR` vignettes:

```{r test-vignettes}
vignette(package = "c14bazAAR")
```

#### **comments:**

The vignette is beautifully written and elegantly and efficiently demonstrates some typical uses of the package.

***

## Test functionality:

- Are there **user interface improvements** that could be made?
- Are there **performance improvements** that could be made?

```{r free-style}
library("c14bazAAR")
```

```{r parse-functions}
exports <-ls("package:c14bazAAR")
exports
```

#### **comments:**

I ran some of the code in the README.md and everything worked as expected. I appreciated the comments that some code would take long to run!

***

## Inspect code:

- Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?
    * good [function & variable naming?](https://ropensci.github.io/dev_guide/building.html#function-and-argument-naming)
    * good [dependency management](https://ropensci.github.io/dev_guide/building.html#package-dependencies)?
- Are there **improvements** that could be made to the [**code style?**](https://ropensci.github.io/dev_guide/building.html#code-style)
- Is there **code duplication** in the package that should be reduced?

```{r inspect-code, eval=FALSE}
pkgreviewr::pkgreview_print_source("c14bazAAR")
```

#### **comments:**

I inspected a sample of functions and everything I saw looked excellent.

One suggestion for improvement: in `get_country_thesaurus` and similar `get_` functions I see we are downloading data, eg. from `"https://raw.githubusercontent.com/ISAAKiel/c14bazAAR/master/data-raw/material_thesaurus.csv"`. Looking in `data-raw/` I see that these files are not large (none more than 100 KB). So I suggest that the authors make these exported data objects in the package so users don't have to download again.

## Review test suite:

See guidance on [testing](https://ropensci.github.io/dev_guide/building.html#testing) for further details.

### test coverage

```{r pkg_coverage}
covr::package_coverage(pkg_dir)

```

### inspect [tests](https://github.com/ISAAKiel/c14bazAAR/blob/master/tests/testthat)

```
c14bazAAR Coverage: 42.53%
R/get_14sea.R: 0.00%
R/get_adrac.R: 0.00%
R/get_austarch.R: 0.00%
R/get_calpal.R: 0.00%
R/get_context.R: 0.00%
R/get_eastafrica.R: 0.00%
R/get_eubar.R: 0.00%
R/get_euroevol.R: 0.00%
R/get_palmisano.R: 0.00%
R/get_radon.R: 0.00%
R/get_radonb.R: 0.00%
R/helpers_db_urls.R: 0.00%
R/zzz.R: 0.00%
R/get_all_dates.R: 40.00%
R/c14_date_list_basic.R: 42.11%
R/helpers_general.R: 62.50%
R/c14_date_list_spatial_determine_country_by_coordinate.R: 71.83%
R/c14_date_list_fuse.R: 81.82%
R/c14_date_list_spatial_finalize_country_name.R: 87.50%
R/c14_date_list_classify_material.R: 89.47%
R/c14_date_list_spatial_standardize_country_name.R: 92.50%
R/c14_date_list_convert.R: 92.86%
R/c14_date_list_duplicates_remove.R: 93.02%
R/c14_date_list_spatial_coordinate_precision.R: 93.10%
R/c14_date_list_duplicates_mark.R: 94.87%
R/c14_date_list_calibrate.R: 95.92%
R/c14_date_list_enforce_types.R: 96.43%
R/c14_date_list_order_variables.R: 97.22%
R/helpers_thesauri.R: 100.00%
```

#### **comments:**

I'm satisfied with the testing coverage.

***
