### 2019-09-09 #333 @ercrema
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is an excellent package that will undoubtedly help many users handling radiocarbon datasets. I tested several commands and had no problems at all. I strongly recommend its publication on rOpenSci. Some minor points below:

* _Statement of Need_ should have perhaps a separate section from Summary in `paper.md`.
*  Oslisly:2013 does not have a DOI
* There is currently a possible error in `BchronCalibrate()` yielding a small discrepancy with the calibration output of other packages such as  `rcarbon`  and `OxCal` (via oxcAAR). I appreciate the authors cannot do much, but thought this was worth mentioning here. There is currently a Pull Request that might fix the problem.
* The functions `coordinate_precision()`, `enforce_types()`, `fuse()` do not have examples. I don't think this is crucial though.
* I wonder whether it makes more sense to have a single `get_c14data` function rather than separate ones for each c14 archive. Or perhaps make these functions internal and have a wrapper calling these functions, perhaps returning a list of available databases when the user does not specify one?
