+++ paper.bib
@@ -166,7 +166,8 @@
   number = {2--3},
   journal = {Radiocarbon},
   author = {Oslisly, Richard and Bentaleb, Ilham and Favier, Charly and Fontugne, Michel and Gillet, Jean Fran{\c c}ois and {Morin-Rivat}, Julie},
-  year = {2013}
+  year = {2013},
+  doi = {10.2458/azu_js_rc.55.16385}
 }
 
 
