+++ R/c14_date_list_enforce_types.R
@@ -17,6 +17,19 @@
 #' @return an object of class c14_date_list
 #' @export
 #'
+#' @examples
+#' # initial situation
+#' ex <- example_c14_date_list
+#' class(ex$c14age)
+#'
+#' # modify variable/column type
+#' ex$c14age <- as.character(ex$c14age)
+#' class(ex$c14age)
+#'
+#' # fix type with enforce_types()
+#' ex <- enforce_types(ex)
+#' class(ex$c14age)
+#'
 #' @rdname enforce_types
 #'
 enforce_types <- function(x, suppress_na_introduced_warnings = TRUE) {
