### 2019-09-10 b21770f
separation of the paper.md text into Background&Code Summary + a small change of the sentence structure in the last paragraph
