+++ paper.md
@@ -24,7 +24,7 @@ date: 12 July 2019
 bibliography: paper.bib 
 ---
 
-# Summary
+# Background
 
 Radiocarbon dating is one of the most important methods for absolute and relative chronological reconstruction of cultural development in prehistoric archaeology [@Taylor:2016]. This is true for the intrasite level, for regional comparisons and also in cases, where processes that have a large spatial and temporal reach are to be investigated. Prominent examples for the latter include the 'neolithization' of Europe or the 'Bantu expansion' in sub-Saharan Africa. They have been extensively analysed with radiocarbon data (e.g. @Ammerman:1971, @Garcin:2018, @Jerardino:2014, @Lemmen:2011, @Oslisly:2013, @Pinhasi:2005, @Russell:2014, @Weninger:2009).
 
@@ -34,12 +34,13 @@ Fortunately, many archaeological institutions and individual authors are sharing
 
 However the entire data basis is currently highly decentralized and lacks basic standardisation. That results in an effective loss of the possible added value that could result from the intersection of data sets in terms of searchability, error checking and further analysis. The creation of a world-wide and centralised database of radiocarbon dates which could solve these issues is not to be expected for the near future.
 
+# Code Summary
+
 ``c14bazAAR`` attempts to tackle the problem at hand by providing an independent interface to access different radiocarbon data sources and make them available for a reproducible research process: from modelling to publication to scientific discourse. It queries multiple openly available ^14^C data archives not behind pay- or login-walls.
 
 The package includes download functions `get_*` (e.g. `get_aDRAC()`, `get_EUROEVOL()`) that acquire the databases from different sources online. They reduce the tables to a set of common variables and store them in a dedicated R S3 class: `c14_date_list`. The `c14_date_list` is based on `tibble::tibble` to integrate well into the R [tidyverse](https://www.tidyverse.org/) ecosystem. It also establishes standardised data types for the most important variables usually defined to describe radiocarbon data.
 
-Beyond the download functions, ``c14bazAAR`` contains a multitude of useful helpers that can be applied to objects of class `c14_date_list`. These include methods for the bulk calibration of radiocarbon dates with the Bchron R package [@Haslett:2008], the removal of duplicates, the estimation of coordinate precision, and the conversion to other useful R data types (e.g. `sf::sf` [@Pebesma:2018]). For the classification of sample material ``c14bazAAR`` provides manually curated reference lists that map the inconsistent attributions in the source databases to a standardized set of material classes. Such a reference list exists as well to fix the country attribution value of dates -- which is especially important in case of missing coordinate information. While another function to determine the source country based on coordinated fails on such dates, the country of origin can be determined if only coordinated are provided.
-
+Beyond the download functions, ``c14bazAAR`` contains a multitude of useful helpers that can be applied to objects of class `c14_date_list`. These include methods for the bulk calibration of radiocarbon dates with the Bchron R package [@Haslett:2008], the removal of duplicates, the estimation of coordinate precision, and the conversion to other useful R data types (e.g. `sf::sf` [@Pebesma:2018]). For the classification of sample material ``c14bazAAR`` provides a manually curated reference list that maps the inconsistent attributions in the source databases to a standardized set of material classes. Such a reference list exists as well to fix the country attribution value of dates -- which is especially important in case of missing coordinate information. Another function to determine the source country based on coordinates fails on such dates.
 
 ``c14bazAAR`` was already and is currently used for data aquisition and preparation of multiple research papers, e.g. @Schmid:2019.
 
