+++ DESCRIPTION
@@ -41,7 +41,15 @@ Authors@R:
              family = "Mueller-Scheessel",
              role = "ctb",
              email = "nils.mueller-scheessel@ufg.uni-kiel.de",
-             comment = c(ORCID = "0000-0001-7992-8722")))
+             comment = c(ORCID = "0000-0001-7992-8722")),
+      person(given = "Ben",
+             family = "Marwick",
+             role = "rev",
+             comment = c(ORCID = "0000-0001-7879-4531")),
+      person(given = "Enrico R.",
+             family = "Crema",
+             role = "rev",
+             comment = c(ORCID = "0000-0001-6727-5138")))
 URL: https://github.com/ISAAKiel/c14bazAAR
 BugReports: https://github.com/ISAAKiel/c14bazAAR/issues
 Depends: R (>= 3.4.0)
