### 2019-09-10 3c2804c
first draft of a simpler interface as suggested by @ercrema
