### 2019-09-10 #333 @nevrome
Thank you for your time and valuable feedback, @benmarwick and @ercrema! And of course thank you for the organisation so far, @melvidoni. I spoke with @dirkseidensticker and we decided to answer your requests in the following way:

@benmarwick
1. This unusual raw data download setup was established on purpose to be able to react fast in case of changes in the location (url) or content (new country names, new materials) of the source databases. A CRAN release can take a relatively long time. With this setup we are a lot more flexible and can make things work again immediately.

@ercrema
1. https://github.com/ISAAKiel/c14bazAAR/commit/b21770f435279cc13c08b9204fd4698dde33619b establishes a separation of a *Background* and a *Code Summary* section. This structure is also used by other JOSS papers. I hope that's what you had in mind?
2. DOI added with https://github.com/ISAAKiel/c14bazAAR/commit/ab42faa1c495a19f5f9372ae2cccf19deada76b5.
3. I guess the relevant PR is this one here: https://github.com/andrewcparnell/Bchron/pull/12, ja? Interesting! I'll keep an eye on it.
4. https://github.com/ISAAKiel/c14bazAAR/commit/b4889e5df9bfb702c969ce12ea7f852767423671 adds some example code for these functions.
5. Both Dirk and I like the idea of simplifying the interface in this way. I wrote a first draft [here](https://github.com/ISAAKiel/c14bazAAR/pull/76) and we will explore it further.

Beyond these changes I also added you to the DESCRIPTION file: https://github.com/ISAAKiel/c14bazAAR/commit/2d2ffc4f0a4af8c8dc8344f1c31b2c014dddb50b. Is your information correct?
