### 2019-09-11 #333 @ercrema
1. Yes precisely.
2. Great!
3. Yes indeed - Of course rcarbon and Oxcal might have got this wrong, but I think the calibration curve is reporting errors in sigma and not sigma^2
4. Great - and they all work fine.
5. Looks great!
