### 2019-09-16 #333 @nevrome
Thanks for these instructions, @melvidoni.

I was not fully aware that the package is intended to be hosted at the "ropenci" github organization at the end of the review. I would rather not do that. c14bazAAR is currently placed at the [ISAAKiel](https://github.com/ISAAKiel) github organization and I would prefer not to move it. ISAAK is a brand with recognition value in the field of computational archaeology.

What are the consequences of this decision and would you advise me to reconsider?
