### 2019-09-16 #333 @benmarwick
If I could suggest one possible option for the best of both worlds: use multiple remotes. Would you consider to mirror the pkg on rOpenSci, and keep the official repo on ISAAK? Clemens can then simultaneously push to both remotes to easily keep them in sync.
