### 2019-09-23 #333 @stefaniebutland
Hi @nevrome. I'm rOpenSci's Community Manager. Once the archiving issue is resolved, would you be interested in writing a blog post about some aspect of c14bazAAR? To the best of my knowledge, this is the first Archaeology-focused package to go through rOpenSci software peer review so a post would garner some attention. @benmarwick said he'd be happy to contribute if you feel that's appropriate.

This tag gets you (diverse) examples of blog posts by authors of peer-reviewed packages: https://ropensci.org/tags/software-peer-review/.

Here are some technical and editorial guidelines: https://github.com/ropensci/roweb2#contributing-a-blog-post. Publication date is flexible. I like to get a draft via pull request a week before the planned publication date so I can review.

Happy to answer any questions.
