### 2019-09-24 #333 @nevrome
@stefaniebutland Ja - absolutely! That's a great idea and I would love to do that.

I'm sure you have some ideas, which aspects of c14bazAAR might be the most interesting for this purpose @benmarwick. I have some as well. Let's discuss this, when the archiving is sorted out.
