### 2019-09-27 d61fa10
experiments with ridge plot mapping
