### 2019-09-28 ebf38c0
more experiments with spatial data
