+++ playground/ridges_map.R
@@ -1,7 +1,7 @@
 library(ggplot2)
 library(ggridges)s
 
-world_wgs84 <- rnaturalearth::ne_coastline(110, "sf")
+world_wgs84 <- rnaturalearth::ne_coastline(50, "sf")
 
 border_wgs84 <-
   sf::st_polygon(list(rbind(
@@ -15,14 +15,15 @@ border_wgs84 <-
   #sf::st_transform(crs = 25832) %>%
   #sf::st_buffer(dist = 0)
 
-border_25832 <- border_wgs84 %>%
-  sf::st_transform(crs = 25832)
-
 europe_wgs84 <- world_wgs84 %>%
   sf::st_crop(border_wgs84)
 
 europe_25832 <- europe_wgs84 %>%
   sf::st_transform(crs = 25832)
+
+st_envelope = function(x) sf::st_as_sfc(sf::st_bbox(x))
+border_europe_25832 <- europe_25832 %>%
+  st_envelope()
 
 plot(europe_25832)
 
@@ -34,7 +35,7 @@ a2 <- a1 %>% dplyr::filter(
 
 a2b <- a2 %>% c14bazAAR::as.sf() %>%
   sf::st_transform(crs = 25832) %>%
-  sf::st_crop(border_25832)
+  sf::st_crop(border_europe_25832)
 
 a3 <- a2b %>%
   dplyr::mutate(
@@ -47,29 +48,40 @@ a3 <- a2b %>%
 
 colnames(a3) <- gsub("data.", "", colnames(a3))
 
-ad <- 1950
-
 a4 <- a3 %>%
   dplyr::group_by(
     X_r, Y_r
   ) %>%
   dplyr::summarise(
-    age = max(c14age) - ad
+    age_max = max(c14age),
+    age_min = min(c14age),
+    age_range = abs(age_max - age_min) * 2,
+    n = dplyr::n() * 200
   ) %>%
   dplyr::ungroup()
 
 a5 <- a4 %>% tidyr::complete(
-  X_r, Y_r, fill = list(age = 0)
+  X_r = tidyr::full_seq(X_r, 10000), Y_r = tidyr::full_seq(Y_r, 10000), fill = list(
+    age_max = 0, age_min = 0, age_range = 0, n = 0
+  )
+)
+
+a6 <- a5 %>% tidyr::pivot_longer(
+  age_max:n,
+  "type"
 )
 
 ggplot() +
   geom_sf(data = europe_25832) +
   geom_ridgeline(
-    data = a5,
-    mapping = aes(x = X_r, y = Y_r, group = Y_r, height = age),
+    data = a6,
+    mapping = aes(x = X_r, y = Y_r, group = Y_r, height = value),
     alpha = 0.3,
-    color = "red",
-    #fill = "red",
-    scale = 7
-  )
+    color = "black",
+    fill = "green",
+    scale = 5,
+    size = 0.05
+  ) +
+  facet_wrap(~type) +
+  theme_bw()
 
