### 2019-10-21 #333 @melvidoni
Hello @nevrome, apologies for the extended delay. The package needs to be under our GitHub repo, but there's no reason it can't also be under another GitHub repo, too, and there are several reasonable solutions for this (multiple remotes, update by CI, etc.). Could you set up something like that?
