### 2019-10-21 #333 @nevrome
@melvidoni The delay was mutual: I just started a new job and was not able to think much about this project. I will do some research about the possible solutions and come back to you asap.
