### 2019-11-06 #333 @nevrome
@melvidoni I discussed this now with the ISAAK people and decided to transfer the package to the ropensci organisation. I will establish a mirror system via CI to keep a copy of the repo at the ISAAKiel organisation.

Unfortunately I was not able to start the transfer: `You don’t have the permission to create public repositories on ropensci`. What's the problem here?
