### 2019-11-06 #333 @melvidoni
Hello @nevrome. You need to first transfer the repo (you need to be part of the rOpenSci org, and _transfer_ not _create_ the repo), and then I'll assign you admin rights on it. That is how we always do it. Is transfer what you tried?
