+++ R/c14_date_list_basic.R
@@ -7,7 +7,7 @@
 #' @description The \strong{c14_date_list} is the central data structure of the
 #' \code{c14bazAAR} package. It's a tibble with set of custom methods and
 #' variables. Please see the
-#' \href{https://github.com/ISAAKiel/c14bazAAR/blob/master/data-raw/variable_reference.csv}{variable_reference}
+#' \href{https://github.com/ropensci/c14bazAAR/blob/master/data-raw/variable_reference.csv}{variable_reference}
 #' table for a description of the variables. Further available variables are ignored. \cr
 #' If an object is of class data.frame or tibble (tbl & tbl_df), it can be
 #' converted to an object of class \strong{c14_date_list}. The only requirement
