+++ man/db_getter.Rd
@@ -50,7 +50,7 @@ current URL of the data source}
 \description{
 This functions downloads source databases and adjust their variables to conform to the
 definition in the
-\href{https://github.com/ISAAKiel/c14bazAAR/blob/master/data-raw/variable_reference.csv}{variable_reference}
+\href{https://github.com/ropensci/c14bazAAR/blob/master/data-raw/variable_reference.csv}{variable_reference}
 table. That includes renaming and arranging the variables (with \code{c14bazAAR::order_variables()})
 as well as type conversion (with \code{c14bazAAR::enforce_types()}) -- so all the steps undertaken by
 \code{as.c14_date_list()}. \cr
