+++ man/c14bazAAR-package.Rd
@@ -11,8 +11,8 @@ Query different C14 date databases and apply basic data cleaning, merging and ca
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://github.com/ISAAKiel/c14bazAAR}
-  \item Report bugs at \url{https://github.com/ISAAKiel/c14bazAAR/issues}
+  \item \url{https://github.com/ropensci/c14bazAAR}
+  \item Report bugs at \url{https://github.com/ropensci/c14bazAAR/issues}
 }
 
 }
