### 2019-11-08 #333 @stefaniebutland
@nevrome Great to hear things are sorting out. I'm keen to host a blog post when you have time to work on this. Both you and @benmarwick are in rOpenSci Slack so we could move the blog post discussion to a DM there.
