### 2019-11-16 95f8151
removed link to neolithicrc
