### 2019-11-16 b24b259
update travis yml to enable automatic update of mirror
