+++ .travis.yml
@@ -1,18 +1,18 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
-
 language: R
 sudo: true
 cache: packages
 r_packages:
-  - covr
-
+- covr
 before_install:
-  - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
-  - sudo apt-get --yes --force-yes update -qq
-  # units/udunits2 dependency:
-  - sudo apt-get install --yes libudunits2-dev
-  # sf dependencies:
-  - sudo apt-get install --yes libproj-dev libgeos-dev libgdal-dev
-
+- sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
+- sudo apt-get --yes --force-yes update -qq
+- sudo apt-get install --yes libudunits2-dev
+- sudo apt-get install --yes libproj-dev libgeos-dev libgdal-dev
 after_success:
-  - Rscript -e 'library(covr); codecov()'
+- Rscript -e 'library(covr); codecov()'
+- git clone --bare https://github.com/ropensci/c14bazAAR.git
+- cd c14bazAAR.git/
+- git push --mirror https://${GH_TOKEN}@github.com/ISAAKiel/c14bazAAR.mirror.git
+env:
+  global:
+    secure: ksgEc/DJzkFDj21M9txQqDaxsZ0RHiDTg1SZwRCE2i88oQjeN1mwENNvqOgTSZAxrRSiDe2r89zjlKTakiVFuXALG+9DOJ+LQCS6C2saqok9/J2/2emPSLJ4HhOlsT+9xsZiYkpVM9HTpiCycqHYL03eql1edps9KWDI02ipWSygBAx/OaDpn/luwHVWrOaULWiOQyvAXmv61RGsmktqqDXds/C9eFKPSI1vk2aw6Lv+8oNFe5OtEQ432mOKlhKdP8JXVxMLAm3Iz35JvP8bw9lRzdaXuoDMemdCxHlJ+NndnFeLqDGmDOM2QuxXNkiXo0t5WpK4yq18yTiJEDsEQWr1PJOvNNn51swxeQIQSHxWrW+2CB77s0po1rKwRE8hOXS+zUeg8F5l4/LhdugFIfL5/MJC/hezwhLm58Rn6n27/IpPkm2pUQjcVf5PTjy2AyCO4v/Zy3rHfinuvNBHfNyr3CU9f7Rk5uXQGou5S7Zzf3nP2F6luYzL/SFGvqHCq1PCtGYcppyMVbjulfaMCLhxgpt/UBk8c8r2/YN1BhEXgx0Is/xCmV8hXglW+fXLPbb/8MqIswgJbAVd6RuDf4nGgBWGjrRmUdQ/TWh3Cp9FA4sRW0e1Gfbrp9nLvFQGjdcdtU2bUJ1P5eKANphtrZLkRWfk4sf6kbVTexP51NQ=
