### 2019-11-16 f625d0c
fixed merge conflict in README
