### 2019-11-16 1a88b5d
added an error message if a wrong name is among the listed databases
