### 2019-11-16 ff9129c
better download error message: now with name of the database that actually failed
