### 2019-11-16 d497345
forgot to consider "all"
