### 2019-11-16 5ea0bd3
increasing robustness -> every input to lower
