### 2019-11-16 b4959d4
check for the special case of no successful download
