### 2019-11-17 25f3056
added RopenSci footer to the bottom of the README
