### 2019-11-17 072f680
fixed some small issues in paper.md
