### 2019-11-17 9e5405e
wrong formatting in paper.bib
