+++ paper.bib
@@ -1,5 +1,5 @@
 @book{Taylor:2016,
-  title = {Radiocarbon Dating} : {An Archaeological Perspective}},
+  title = {Radiocarbon Dating: An Archaeological Perspective},
   isbn = {978-1-315-42121-6},
   shorttitle = {Radiocarbon Dating},
   language = {en},
@@ -145,7 +145,6 @@
   doi = {10.1371/journal.pone.0087854}
 }
 
-
 @article{Jerardino:2014,
   title = {Cultural {{Diffusion Was}} the {{Main Driving Mechanism}} of the {{Neolithic Transition}} in {{Southern Africa}}},
   volume = {9},
@@ -159,7 +158,6 @@
   pages = {1--19}
 }
 
-
 @article{Oslisly:2013,
   title = {West {{Central African Peoples}}: {{Survey}} of {{Radiocarbon Dates}} over the {{Past}} 5000 {{Years}}},
   volume = {55},
@@ -169,7 +167,6 @@
   year = {2013},
   doi = {10.2458/azu_js_rc.55.16385}
 }
-
 
 @article{Garcin:2018,
   title = {Early Anthropogenic Impact on {{Western Central African}} Rainforests 2,600 y Ago},
