### 2019-11-17 #333 @nevrome
@melvidoni: I fulfilled all the ToDos in https://github.com/ropensci/software-review/issues/333#issuecomment-531599658 now. Except for the blog post.
