### 2019-11-18 #333 @arfon
> - [ ] Submit to JOSS using the Zenodo DOI. They will tag it for expedited review.

:wave: @melvidoni - this isn't quite correct sorry, we actually need authors to submit to JOSS using the repository address (https://github.com/ropensci/c14bazAAR). Is there an rOpenSci template I can propose a modification to?
