### 2019-11-18 #333 @melvidoni
> > * [ ]  Submit to JOSS using the Zenodo DOI. They will tag it for expedited review.
>
> 👋 @melvidoni - this isn't quite correct sorry, we actually need authors to submit to JOSS using the repository address (https://github.com/ropensci/c14bazAAR). Is there a rOpenSci template I can propose a modification to?

Yes, there is. Let me mention this to the other editors first, please.
