### 2019-11-18 fa02feb
integrated changes suggested by @dirkseidensticker and @MartinHinz
