### 2019-11-18 327243a
fixed some details in the bibliography
