+++ paper.bib
@@ -17,16 +17,14 @@
   number = {4},
   pages = {674--688},
   publisher = {Wiley, Royal Anthropological Institute of Great Britain and Ireland},
-  title = {Measuring the Rate of Spread of Early Farming in Europe},
+  title = {Measuring the Rate of Spread of Early Farming in {Europe}},
   volume = {6},
   year = 1971,
   doi = {10.2307/2799190}
 }
 
 @misc{Palmisano:2017,
-  title = {Regional {{Demographic Trends}} and {{Settlement Patterns}} in {{Central Italy}}: {{Archaeological Sites}} and {{Radiocarbon Dates}}},
-  copyright = {open},
-  shorttitle = {Regional {{Demographic Trends}} and {{Settlement Patterns}} in {{Central Italy}}},
+  title = {Regional Demographic Trends and Settlement Patterns in {Central} {Italy}: Archaeological Sites and Radiocarbon Dates},
   journal = {UCL Institute of Archaeology: London, UK. (2017)},
   author = {Palmisano, A. and Bevan, A. and Shennan, S.},
   month = sep,
@@ -35,7 +33,7 @@
 }
 
 @article{Haslett:2008,
-  title = {A simple monotone process with application to radiocarbon-dated depth chronologiesy},
+  title = {A simple monotone process with application to radiocarbon-dated depth chronologies},
   author = {John Haslett and Andrew C Parnell},
   journal = {Journal of the Royal Statistical Society: Series C (Applied Statistics)},
   volume = {57},
@@ -48,7 +46,7 @@
 
 @article{Pebesma:2018,
   author = {Edzer Pebesma},
-  title = {Simple Features for R: Standardized Support for Spatial Vector Data},
+  title = {Simple Features for {R}: Standardized Support for Spatial Vector Data},
   journal = {{The R Journal}},
   pages = {439--446},
   volume = {10},
@@ -58,8 +56,8 @@
 }
 
 @misc{Kneisel:2013,
-  title = {Radon-B},
-  url = {http://radon-b.ufg.uni-kiel.de.},
+  title = {{Radon-B}},
+  howpublished = {\url{http://radon-b.ufg.uni-kiel.de}},
   author = {Kneisel, Jutta and Hinz, Martin and Rinne, Christoph},
   year = 2013
 }
@@ -67,19 +65,18 @@
 @data{Mustaphi:2016,
   author = {Courtney Mustaphi, Colin},
   publisher = {Harvard Dataverse},
-  title = {Radiocarbon dates from eastern Africa in the CARD2.0 format},
+  title = {Radiocarbon dates from {Eastern} {Africa} in the {CARD2.0} format},
   year = 2016,
   doi = {10.7910/DVN/NJLNRJ}
 }
 
 @article{Schmid:2019,
   author = {Schmid, Clemens},
-  title = {Evaluating Cultural Transmission in Bronze Age burial rites of Central, Northern and Northwestern Europe using radiocarbon data},
+  title = {Evaluating Cultural Transmission in {Bronze} {Age} burial rites of {Central}, {Northern} and {Northwestern} {Europe} using radiocarbon data},
   journal = {Adaptive Behavior},
   year = 2019,
   doi = {10.1177/1059712319860842}
 }
-
 
 @article{Hinz:2012,
 	title = {{RADON} - {Radiocarbon} dates online 2012. {Central} {European} database of 14C dates for the {Neolithic} and the {Early} {Bronze} {Age}.},
@@ -94,7 +91,7 @@
 }
 
 @article{Weninger:2009,
-  title = {The Impact of Rapid Climate Change on Prehistoric Societies during the Holocene in the Eastern Mediterranean},
[…]
