+++ paper.bib
@@ -5,7 +5,6 @@
   language = {en},
   publisher = {Routledge},
   author = {Taylor, R. E. and {Bar-Yosef}, Ofer},
-  month = jun,
   year = 2016,
   doi = {10.4324/9781315421216}
 }
@@ -27,7 +26,6 @@
   title = {Regional Demographic Trends and Settlement Patterns in {Central} {Italy}: Archaeological Sites and Radiocarbon Dates},
   journal = {UCL Institute of Archaeology: London, UK. (2017)},
   author = {Palmisano, A. and Bevan, A. and Shennan, S.},
-  month = sep,
   year = 2017,
   doi = {10.14324/000.ds.1575442}
 }
@@ -86,7 +84,6 @@
 	doi = {10.12766/jna.2012.65},
 	journal = {Journal of Neolithic Archaeology},
 	author = {Hinz, Martin and Furholt, Martin and Müller, Johannes and Rinne, Christoph and Raetzel-Fabian, Dirk and Sjögren, Karl-Göran and Wotzka, Hans-Peter},
-	month = feb,
 	year = 2012
 }
 
@@ -98,7 +95,6 @@
   journal = {Documenta Praehistorica},
   author = {Weninger, Bernhard and Clare, Lee and Rohling, Eelco and Bar-Yosef, Ofer and Böhner, Utz and Budja, Mihael and Bundschuh, Manfred and Feurdean, Angelica and Gebe, Hans Georg and Jöris, Olaf and Linstädter, Jörg and Mayewski, Paul and Mühlenbruch, Tobias and Reingruber, Agathe and Rollefson, Gary and Schyle, Daniel and Thissen, Laurens and Todorova, Henrieta and Zielhofer, Christoph},
   year = 2009,
-  month = dec,
   pages = {7--59}
 }
 
@@ -120,7 +116,6 @@
   publisher = {Public Library of Science},
   title = {Tracing the Origin and Spread of Agriculture in {Europe}},
   year = {2005},
-  month = {11},
   volume = {3},
   number = {12},
   doi = {10.1371/journal.pbio.0030410}
@@ -132,7 +127,6 @@
   publisher = {Public Library of Science},
   title = {Modelling the Spread of Farming in the Bantu-Speaking Regions of {Africa}: An Archaeology-Based Phylogeography},
   year = {2014},
-  month = {01},
   volume = {9},
   pages = {1--9},
   number = {1},
