### 2019-11-24 1c88e42
started to work on IR-DD parser
