### 2019-11-25 d7198b0
first version of get_IRDD
