### 2019-11-25 c29b994
moved get_irdd to R/
