### 2019-11-27 caabcb7
replaced hard coded URLs with function arguments to get helpers
