+++ man/get_db_url.Rd
@@ -4,7 +4,11 @@
 \alias{get_db_url}
 \title{get db url}
 \usage{
-get_db_url(db_name)
+get_db_url(
+  db_name,
+ 
+    ref_url = "https://raw.githubusercontent.com/ropensci/c14bazAAR/master/data-raw/url_reference.csv"
+)
 }
 \arguments{
 \item{db_name}{name of the database}
