### 2019-11-27 0f95452
update of NEWs.md
