+++ R/helpers_db_urls.R
@@ -4,6 +4,7 @@
 #' on github.
 #'
 #' @param db_name name of the database
+#' @param ref_url url of the relevant reference table
 #'
 #' @export
 get_db_url <- function(db_name, ref_url = "https://raw.githubusercontent.com/ropensci/c14bazAAR/master/data-raw/url_reference.csv") {
