+++ R/helpers_thesauri.R
@@ -5,10 +5,16 @@
 #' @param ref_url url of the relevant reference table
 #'
 #' @export
-get_country_thesaurus <- function(ref_url = "https://raw.githubusercontent.com/ropensci/c14bazAAR/master/data-raw/country_thesaurus.csv") {
-  ref_url %>%
-    get_thesaurus() %>%
-    return()
+get_country_thesaurus <- function(
+  ref_url = paste(c(
+    "https://raw.githubusercontent.com",
+    "ropensci",
+    "c14bazAAR",
+    "master",
+    "data-raw",
+    "country_thesaurus.csv"
+  ), collapse = "/")) {
+  ref_url %>% get_thesaurus() %>% return()
 }
 
 #' get_material_thesaurus
@@ -18,11 +24,17 @@ get_country_thesaurus <- function(ref_url = "https://raw.githubusercontent.com/r
 #' @param ref_url url of the relevant reference table
 #'
 #' @export
-get_material_thesaurus <- function(ref_url = "https://raw.githubusercontent.com/ropensci/c14bazAAR/master/data-raw/material_thesaurus.csv") {
-  ref_url %>%
-    get_thesaurus() %>%
-    return()
-}
+get_material_thesaurus <- function(
+  ref_url = paste(c(
+    "https://raw.githubusercontent.com",
+    "ropensci",
+    "c14bazAAR",
+    "master",
+    "data-raw",
+    "material_thesaurus.csv"
+  ), collapse = "/")) {
+    ref_url %>% get_thesaurus() %>% return()
+  }
 
 #' get_thesaurus
 #'
