### 2019-11-30 a1275a6
first logo draft
