### 2019-11-30 26c38f7
experiments with logo position
