+++ README.md
@@ -1,3 +1,7 @@
+<p align="center">
+  <img src="inst/logo.png" width = 200>
+</p>
+
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![Travis-CI Build Status](https://travis-ci.org/ropensci/c14bazAAR.svg?branch=master)](https://travis-ci.org/ropensci/c14bazAAR) [![Coverage Status](https://img.shields.io/codecov/c/github/ropensci/c14bazAAR/master.svg)](https://codecov.io/github/ropensci/c14bazAAR?branch=master)
 [![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/c14bazAAR)](https://cran.r-project.org/package=c14bazAAR)
@@ -5,11 +9,9 @@
 [![DOI](https://img.shields.io/badge/DOI-10.17605%2FOSF.IO%2F3DS6A-blue)](https://doi.org/10.17605/OSF.IO/3DS6A)
 [![DOI](https://joss.theoj.org/papers/10.21105/joss.01914/status.svg)](https://doi.org/10.21105/joss.01914)
 
-# c14bazAAR
-
-<img align="right" src="inst/image/khajiit.jpg" width = 430>
-
 c14bazAAR is a R package to query different openly accessible radiocarbon date databases. It allows basic data cleaning, calibration and merging. If you're not familiar with R other tools (such as [GoGet](https://www.ibercrono.org/goget/index.php)) to search for radiocarbon dates might be better suited for your needs.
+
+
 
 - [Installation](#installation)
 - [How to use](#how-to-use)
