+++ README.md
@@ -13,22 +13,12 @@ c14bazAAR is a R package to query different openly accessible radiocarbon date d
 
 
 
-- [Installation](#installation)
-- [How to use](#how-to-use)
-  - [Download](#download)
-  - [Calibration](#calibration)
-  - [Material classification](#material-classification)
-  - [Country attribution](#country-attribution)
-  - [Duplicates](#duplicates)
-  - [Coordinate precision](#coordinate-precision)
-  - [Conversion](#conversion)
-  - [Technical functions](#technical-functions)
-  - [Plotting and visualization](#plotting-radiocarbon-data)
-- [Databases](#databases)
-- [Contributing](#contributing)
-  - [Adding database getter functions](#adding-database-getter-functions)
-- [Citation](#citation)
-- [License](#license)
+- [**Installation**](#installation)
+- [**How to use**](#how-to-use) ([Download](#download), [Calibration](#calibration), [Material classification](#material-classification), [Country attribution](#country-attribution), [Duplicates](#duplicates), [Coordinate precision](#coordinate-precision), [Conversion](#conversion), [Technical functions](#technical-functions), [Plotting and visualization](#plotting-radiocarbon-data))
+- [**Databases**](#databases)
+- [**Contributing**](#contributing) ([Adding database getter functions](#adding-database-getter-functions))
+- [**Citation**](#citation)
+- [**License**](#license)
 
 If you want to use data downloaded with c14bazAAR or neolithicRC for your research, you have to cite the source databases. Most databases have a preferred way of citation that also may change over time with new versions and publications. Please check the [respective homepages](#databases) to find out more. The output of c14bazAAR does not contain the full citations of the individual dates, but only a short reference tag. For further information you have to consult the source databases.
 
