### 2019-12-08 f5d9776
experiments with a simple plotting option for c14bazAAR
