### 2019-12-08 3ba0022
added plot function
