+++ man/c14_date_list.Rd
@@ -6,6 +6,7 @@
 \alias{is.c14_date_list}
 \alias{format.c14_date_list}
 \alias{print.c14_date_list}
+\alias{plot.c14_date_list}
 \title{\strong{c14_date_list}}
 \usage{
 as.c14_date_list(x, ...)
@@ -15,6 +16,8 @@ is.c14_date_list(x, ...)
 \method{format}{c14_date_list}(x, ...)
 
 \method{print}{c14_date_list}(x, ...)
+
+\method{plot}{c14_date_list}(x, ...)
 }
 \arguments{
 \item{x}{an object}
