### 2019-12-09 7024337
Add links to official docs site. For more details: https://ropensci.org/technotes/2019/06/07/ropensci-docs/
