+++ DESCRIPTION
@@ -50,7 +50,7 @@ Authors@R:
              family = "Crema",
              role = "rev",
              comment = c(ORCID = "0000-0001-6727-5138")))
-URL: https://github.com/ropensci/c14bazAAR
+URL: https://docs.ropensci.org/c14bazAAR, https://github.com/ropensci/c14bazAAR
 BugReports: https://github.com/ropensci/c14bazAAR/issues
 Depends: R (>= 3.4.0)
 Language: en_GB
