### 2019-12-09 544a046
ldraft for logo  with streets in the background
