### 2019-12-09 f38ea6c
update of citation
