### 2019-12-09 db56391
first draft of a write function to solve #39
