+++ NAMESPACE
@@ -26,6 +26,8 @@ S3method(remove_duplicates,c14_date_list)
 S3method(remove_duplicates,default)
 S3method(standardize_country_name,c14_date_list)
 S3method(standardize_country_name,default)
+S3method(write_c14,c14_date_list)
+S3method(write_c14,default)
 export(as.c14_date_list)
 export(as.sf)
 export(calibrate)
@@ -57,6 +59,7 @@ export(mark_duplicates)
 export(order_variables)
 export(remove_duplicates)
 export(standardize_country_name)
+export(write_c14)
 importFrom(magrittr,"%<>%")
 importFrom(magrittr,"%>%")
 importFrom(rlang,":=")
