### 2019-12-15 0ffaf6b
roxygen updates
