### 2019-12-15 e27c922
changed url info getter to also work with version info
