+++ man/get_db_info.Rd
@@ -1,19 +1,21 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/helpers_db_urls.R
-\name{get_db_url}
-\alias{get_db_url}
-\title{get db url}
+% Please edit documentation in R/helpers_dbs.R
+\name{get_db_info}
+\alias{get_db_info}
+\title{get db info}
 \usage{
-get_db_url(db_name,
+get_db_info(db_name, info_type = c("url", "version"),
   ref_url = paste(c("https://raw.githubusercontent.com", "ropensci",
   "c14bazAAR", "master", "data-raw", "url_reference.csv"), collapse = "/"))
 }
 \arguments{
 \item{db_name}{name of the database}
 
+\item{info_type}{type of information: "url", "version"}
+
 \item{ref_url}{url of the relevant reference table}
 }
 \description{
-Downloads URLs of c14 source databases from a reference table
+Downloads information for c14 source databases from a reference table
 on github.
 }
