### 2019-12-15 1a7fc73
added wrappers get_db_url and get_db_version
