+++ man/get_db_version.Rd
@@ -1,17 +1,15 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/helpers_dbs.R
-\name{get_db_info}
-\alias{get_db_info}
-\title{get db info}
+\name{get_db_version}
+\alias{get_db_version}
+\title{get db version}
 \usage{
-get_db_info(db_name, info_type = c("url", "version"),
+get_db_version(db_name,
   ref_url = paste(c("https://raw.githubusercontent.com", "ropensci",
   "c14bazAAR", "master", "data-raw", "url_reference.csv"), collapse = "/"))
 }
 \arguments{
 \item{db_name}{name of the database}
-
-\item{info_type}{type of information: "url", "version"}
 
 \item{ref_url}{url of the relevant reference table}
 }
