+++ NAMESPACE
@@ -50,7 +50,8 @@ export(get_aDRAC)
 export(get_all_dates)
 export(get_c14data)
 export(get_country_thesaurus)
-export(get_db_info)
+export(get_db_url)
+export(get_db_version)
 export(get_material_thesaurus)
 export(is.c14_date_list)
 export(mark_duplicates)
