### 2019-12-15 5d820ab
added version column to individual database getters
