+++ man/get_material_thesaurus.Rd
@@ -4,10 +4,10 @@
 \alias{get_material_thesaurus}
 \title{get_material_thesaurus}
 \usage{
-
-  get_material_thesaurus(ref_url = paste(c("https://raw.githubusercontent.com",
-  "ropensci", "c14bazAAR", "master", "data-raw", "material_thesaurus.csv"),
-  collapse = "/"))
+get_material_thesaurus(
+  ref_url = paste(c("https://raw.githubusercontent.com", "ropensci", "c14bazAAR",
+    "master", "data-raw", "material_thesaurus.csv"), collapse = "/")
+)
 }
 \arguments{
 \item{ref_url}{url of the relevant reference table}
