+++ man/get_country_thesaurus.Rd
@@ -4,10 +4,10 @@
 \alias{get_country_thesaurus}
 \title{get_country_thesaurus}
 \usage{
-
-  get_country_thesaurus(ref_url = paste(c("https://raw.githubusercontent.com",
-  "ropensci", "c14bazAAR", "master", "data-raw", "country_thesaurus.csv"),
-  collapse = "/"))
+get_country_thesaurus(
+  ref_url = paste(c("https://raw.githubusercontent.com", "ropensci", "c14bazAAR",
+    "master", "data-raw", "country_thesaurus.csv"), collapse = "/")
+)
 }
 \arguments{
 \item{ref_url}{url of the relevant reference table}
