### 2019-12-15 c96ab7c
reactivated url check
