+++ R/helpers_dbs.R
@@ -59,7 +59,7 @@ get_db_info <- function(
     "url_reference.csv"
   ), collapse = "/")) {
 
-  #check_connection_to_url(ref_url)
+  check_connection_to_url(ref_url)
 
   if (length(db_name) > 1) {
     stop("get_db_info only works for one database at a time")
