+++ R/helpers_dbs.R
@@ -11,7 +11,7 @@ get_db_url <- function(db_name, ref_url = paste(c(
   "https://raw.githubusercontent.com",
   "ropensci",
   "c14bazAAR",
-  "master",
+  "version",
   "data-raw",
   "url_reference.csv"
   ), collapse = "/")) {
