### 2019-12-28 494b374
update of NEWS.md
