+++ man/duplicates.Rd
@@ -27,8 +27,8 @@ remove_duplicates(x, preferences = NULL, supermerge = FALSE, log = TRUE)
 \item{x}{an object of class c14_date_list}
 
 \item{preferences}{character vector with the order of source databases by
-which the deduping should be executed. If e.g. preferences = c("RADON", "CALPAL")
-and a certain date appears in Radon and EUROEVOL, then only the RADON entry remains.
+which the deduping should be executed. If e.g. preferences = c("radon", "calpal")
+and a certain date appears in radon and euroevol, then only the radon entry remains.
 Default: NULL. With preferences = NULL all overlapping, conflicting information in
 individual columns of one duplicated date is removed. See Option 2 and 3.}
 
