### 2019-12-29 4e1eb05
documented change in NEWS.md
