### 2019-12-29 4ad1748
tests for write_c14 function
