### 2019-12-29 3dbaf36
fixed kiteeastafrica download link
