### 2019-12-29 e5b6b3a
persistent url can not be used for connection check
