### 2019-12-29 1b0c16f
experiments to fix the ca certificate issue with discovery.ucl.ac.uk
