### 2019-12-29 ae9926c
plot needs graphics package
