+++ DESCRIPTION
@@ -61,6 +61,7 @@ Imports:
     crayon (>= 1.3.4),
     data.table (>= 1.11.4),
     dplyr (>= 0.7.2),
+    graphics,
     httr (>= 1.4.0),
     magrittr (>= 1.5),
     pbapply (>= 1.3-3),
