+++ .travis.yml
@@ -12,8 +12,7 @@ before_install:
 - sudo apt-get install --yes libudunits2-dev
 - sudo apt-get install --yes libproj-dev libgeos-dev libgdal-dev
 after_success:
-- Rscript -e "library(c14bazAAR); get_palmisano()"
-- Rscript -e "library(c14bazAAR); get_euroevol()"
+- Rscript -e "devtools::install_github('ropensci/c14bazAAR'); library(c14bazAAR); get_euroevol(); get_palmisano()"
 - Rscript -e 'library(covr); codecov()'
 - git clone --bare https://github.com/ropensci/c14bazAAR.git
 - cd c14bazAAR.git/
