### 2019-12-29 42b4c40
removing travis debugging code
