+++ .travis.yml
@@ -5,7 +5,7 @@ r_packages:
 - covr
 before_install:
 # ca certificates for palmisano and euroevol
-- sudo sh -c "openssl s_client -showcerts -servername discovery.ucl.ac.uk -connect discovery.ucl.ac.uk:443 </dev/null | sed -ne '/-BEGIN CERTIFICATE-/,/-END CERTIFICATE-/p' >> /etc/ssl/certs/ca-certificates.crt"
+- sudo sh -c "wget --no-check-certificate https://www.quovadisglobal.com/~/media/Files/Roots/qvevsslg3.ashx && sudo openssl x509 -inform der -in qvevsslg3.ashx -out /usr/local/share/ca-certificates/qvevsslg3.crt && sudo update-ca-certificates && rm qvevsslg3.ashx"
 # install packages
 - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
 - sudo apt-get --yes --force-yes update -qq
