### 2019-12-30 54eea82
test list column removal
