### 2019-12-30 43e87c6
check some of the helper functions
