### 2019-12-31 005eb6f
clean hexsticker-code & fixed coords
