### 2020-01-01 6eba122
added a white rectangle to increase readability of text
