+++ logo/hexsticker.R
@@ -23,15 +23,25 @@ p <- ggplot() +
                aes(label = label),
                family = 'fontawesome-webfont',
                size = 10)  +
+  geom_rect(
+    aes(
+      xmin = 28.96755,
+      xmax = 28.96945,
+      ymax = 41.01058,
+      ymin = 41.0102,
+    ),
+    fill = "white",
+    alpha = 0.7
+  ) +
   theme_void()
 
 hexSticker::sticker(subplot = p,
                     package="c14bazAAR",
                     p_color = "black",
-                    p_size = 24,
+                    p_size = 20,
                     p_family = "sans",
                     p_x = 1,
-                    p_y = .75,
+                    p_y = .73,
                     s_x = 1,
                     s_y = 1,
                     s_width = 2,
