### 2020-01-01 c986bc0
update of NEWS.md
